fix(test): move badge tests out of forward-ref test file
